reorganize code logically
