+++ .gitignore
@@ -1,3 +1,8 @@
 venv/
 __pycache__/
 .DS_Store
+
+*.jpg
+*.jpeg
+*.png
+*.webp
